Fix on OSM driver. Remove the whole tables Add a unit test

+++ data-manager/src/main/java/org/orbisgis/datamanager/io/IOMethods.java
@@ -125,6 +125,7 @@ public class IOMethods {
             }
             else if (FileUtil.isExtensionWellFormated(fileToImport, "geojson")) {
                 dataSource.execute("DROP TABLE IF EXISTS " + tableName);
+                delete=false;
                 GeoJsonReaderDriver driver = new GeoJsonReaderDriver(connection, fileToImport);
                 driver.read(new EmptyProgressVisitor(), tableName);
                 return true;
@@ -140,7 +141,6 @@ public class IOMethods {
                     LOGGER.warn("Encoding is not yet supported for this file format");
                 }
                 encoding = null;
-                delete = false;
                 driverFunction = new TSVDriverFunction();
             }
             else if (FileUtil.isExtensionWellFormated(fileToImport, "osm") ||
@@ -151,6 +151,7 @@ public class IOMethods {
                 }
                 encoding = null;
                 driverFunction = new OSMDriverFunction();
+
             }
             else if (FileUtil.isExtensionWellFormated(fileToImport, "gpx")) {
                 if( encoding != null && !encoding.isEmpty()) {
@@ -163,7 +164,9 @@ public class IOMethods {
                 LOGGER.error("Unsupported file format");
             }
             if(driverFunction != null){
-                dataSource.execute("DROP TABLE IF EXISTS " + tableName);
+                if(!(driverFunction instanceof OSMDriverFunction)) {
+                    dataSource.execute("DROP TABLE IF EXISTS " + tableName);
+                }
                 if(encoding != null) {
                     driverFunction.importFile(connection, tableName, fileToImport, new EmptyProgressVisitor(), encoding);
                 }
